Moved Log Out Component to Auth Folder
- Moved Log Out component to Auth Folder to clarify the component's functionality
